pyproject.toml
[project]
name = "pm-arbitrage"
version = "0.1.0"
description = "Polymarket-Opinion cross-platform arbitrage system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "py-clob-client>=0.15.0",
    "opinion-clob-sdk>=0.4.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "aiolimiter>=1.1.0",
]

[project.scripts]
arbitrage = "src.main:main"
